Attempt at a cleaner, more maintainable structure
