+++ ota_university/ctf_framework/admin.py
@@ -2,7 +2,6 @@ from django.contrib import admin
 from .models import UserProfile, Title, Challenge, ChallengeCategory
 
 # Register your models here.
-
 admin.site.register(UserProfile)
 admin.site.register(Title)
 admin.site.register(Challenge)
