+++ ota_university/ctf_framework/models.py
@@ -79,14 +79,3 @@ class UserProfile(models.Model):
                 completed_challenges[category.category].append(challenge)
 
         return completed_challenges
-
-
-
-
-
-
-
-
-
-
-
