+++ ota_university/ctf_framework/templates/profile/show.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "../layouts/default.html" %}
 {% block content %}
 
 User: {{ user }}<br>
