+++ ota_university/ctf_framework/views/__init__.py
@@ -1,0 +1,3 @@
+from .home import *
+from .profile import *
+from .challenge import *
