+++ ota_university/ctf_framework/views/home.py
@@ -1,0 +1,13 @@
+from .base_view import *
+def index(request):
+    users = UserProfile.objects.all()
+    challenges = Challenge.objects.all()
+
+    context = {
+        "users": users,
+        "challenges": challenges
+    }
+
+    return render(request, "misc/home.html", context=context)
+
+
